refactor(contracts): make ExecResult host-command vocabulary, not root exec internals
First step of the platform-seam work (#1490 W3 prerequisite / #1478 P6a).

13 of the ~22 modules naming `ExecResult` live under `src/platforms/`, and
every one of them wanted only the type: the shape of a finished host
command, with no runtime behavior at all. Because the type was declared
beside the implementation in `src/utils/exec.ts`, each of those modules
imported root code to describe a value it already held — a dependency no
platform package could ever satisfy, for nothing.

`ExecResult` moves to `@agent-device/contracts/platform`, beside
`PlatformPlugin` and the rest of the platform-family vocabulary. The
running stays in `src/utils/exec.ts`: process spawning, argument policy,
timeouts and redaction are host policy the root composition owns, and
#1490 keeps `exec` root-side deliberately. This splits the vocabulary from
the policy rather than relocating the policy.

Type-only change, no re-export shim at the old path (#1490's rule): every
consumer now names the contracts subpath, and `exec.ts` imports the type it
implements against.

What this does NOT do, stated so the next step is not mistaken for done: 45
platform modules still import `utils/exec.ts` for BEHAVIOR — runCmd (98
references), requireExecSuccess (45), withCommandExecutorOverride (32),
execFailureDetails (30), whichCmd (24), runCmdBackground (21) and a
long tail. That is genuine capability coupling and needs the narrow
injected-capability treatment W1c used for limrun, not a type move.

Gates: typecheck, lint, format, check:layering, and check:affected all
green. Layering unmoved — R6 at 7, R9 at 47, R11 at 10 packages / 32
subpaths.
