+++ src/daemon/__tests__/app-log.test.ts
@@ -1,3 +1,4 @@
+import type { ExecResult } from '@agent-device/contracts/platform';
 import { test } from 'vitest';
 import assert from 'node:assert/strict';
 import fs from 'node:fs';
@@ -6,7 +7,7 @@ import { finished } from 'node:stream/promises';
 import type { DeviceInfo } from '@agent-device/kernel/device';
 import { AppError } from '@agent-device/kernel/errors';
 import { withAppleToolProvider } from '../../platforms/apple/core/tool-provider.ts';
-import type { ExecResult } from '../../utils/exec.ts';
+
 import { runAppLogDoctor, rotateAppLogIfNeeded } from '../app-log.ts';
 import { assertAndroidPackageArgSafe } from '../app-log-android.ts';
 import {
