+++ src/daemon/app-log-process.ts
@@ -1,8 +1,8 @@
+import type { ExecResult } from '@agent-device/contracts/platform';
 import fs from 'node:fs';
 import path from 'node:path';
 import { readProcessCommand, readProcessStartTime } from '../utils/host-process.ts';
 import type { LogBackend } from '@agent-device/contracts/observability';
-import type { ExecResult } from '../utils/exec.ts';
 
 export const APP_LOG_PID_FILENAME = 'app-log.pid';
