+++ src/daemon/recording-provider.ts
@@ -1,6 +1,7 @@
+import type { ExecResult } from '@agent-device/contracts/platform';
 import { buildSimctlArgsForDevice } from '../platforms/apple/core/simctl.ts';
 import type { DeviceInfo } from '@agent-device/kernel/device';
-import { runCmdBackground, type ExecBackgroundResult, type ExecResult } from '../utils/exec.ts';
+import { runCmdBackground, type ExecBackgroundResult } from '../utils/exec.ts';
 import { createScopedProvider } from '../utils/scoped-provider.ts';
 
 export type RecordingProcess = {
