+++ src/platforms/apple/core/__tests__/runner-command-recovery.test.ts
@@ -1,8 +1,9 @@
+import type { ExecResult } from '@agent-device/contracts/platform';
 import assert from 'node:assert/strict';
 import { afterEach, test, vi } from 'vitest';
 import { AppError } from '@agent-device/kernel/errors';
 import { IOS_SIMULATOR } from '../../../../__tests__/test-utils/device-fixtures.ts';
-import type { ExecResult } from '../../../../utils/exec.ts';
+
 import { handleRunnerTransportErrorAfterCommandSend } from '../runner/runner-command-recovery.ts';
 import type { RunnerCommand } from '../runner/runner-contract.ts';
 import type { RunnerSession } from '../runner/runner-session.ts';
