+++ src/platforms/apple/core/runner/runner-adoption.ts
@@ -1,10 +1,11 @@
+import type { ExecResult } from '@agent-device/contracts/platform';
 import path from 'node:path';
 import { resolveIosSimulatorDeviceSetPath } from '../../../../utils/device-isolation.ts';
 import { emitDiagnostic } from '../../../../utils/diagnostics.ts';
 import { isProcessAlive } from '../../../../utils/host-process.ts';
 import { parseBooleanLiteral } from '../../../../utils/source-value.ts';
 import type { DeviceInfo } from '@agent-device/kernel/device';
-import type { ExecResult } from '../../../../utils/exec.ts';
+
 import { sendRunnerCommandOnce } from './runner-transport.ts';
 import { withRunnerCommandId } from './runner-contract.ts';
 import {
