+++ src/platforms/apple/core/runner/runner-session-types.ts
@@ -1,5 +1,5 @@
-import type { RunnerLogicalLeaseContext } from '@agent-device/contracts/platform';
-import type { ExecResult } from '../../../../utils/exec.ts';
+import type { ExecResult, RunnerLogicalLeaseContext } from '@agent-device/contracts/platform';
+
 import type { DeviceInfo } from '@agent-device/kernel/device';
 import type { RunnerXctestrunArtifact } from './runner-xctestrun.ts';
 import type { RunnerLease } from './runner-lease.ts';
