+++ src/platforms/apple/core/tool-provider-types.ts
@@ -1,5 +1,6 @@
+import type { ExecResult } from '@agent-device/contracts/platform';
 import type { AppsFilter } from '@agent-device/contracts/device';
-import type { ExecOptions, ExecResult } from '../../../utils/exec.ts';
+import type { ExecOptions } from '../../../utils/exec.ts';
 import type { IosAppInfo } from './app-info.ts';
 
 export type AppleToolCommandExecutor = (
