+++ src/platforms/linux/tool-provider.ts
@@ -1,7 +1,8 @@
+import type { ExecResult } from '@agent-device/contracts/platform';
 import type { ClickButton, ScrollDirection } from '@agent-device/contracts/interaction';
 import { AppError } from '@agent-device/kernel/errors';
 import { emitDiagnostic } from '../../utils/diagnostics.ts';
-import { runCmd, whichCmd, type ExecOptions, type ExecResult } from '../../utils/exec.ts';
+import { runCmd, whichCmd, type ExecOptions } from '../../utils/exec.ts';
 import { createScopedProvider } from '../../utils/scoped-provider.ts';
 import { sleep } from '../../utils/timeouts.ts';
 import type {
