+++ src/platforms/vega/tool-provider.ts
@@ -1,9 +1,10 @@
+import type { ExecResult } from '@agent-device/contracts/platform';
 import { constants } from 'node:fs';
 import { access } from 'node:fs/promises';
 import os from 'node:os';
 import path from 'node:path';
 import type { VegaTvRemoteKey } from '@agent-device/contracts/interaction';
-import { runCmd, whichCmd, type ExecOptions, type ExecResult } from '../../utils/exec.ts';
+import { runCmd, whichCmd, type ExecOptions } from '../../utils/exec.ts';
 import { createScopedProvider } from '../../utils/scoped-provider.ts';
 
 export type VegaToolProvider = {
